Add initial taglib support for FLAC and MP3 files

+++ Sources/Luminescence.Audio/ChromaprintFingerprinter.cpp
@@ -8,7 +8,8 @@ namespace Luminescence
 {
    namespace Audio 
    {
-      //https://bitbucket.org/acoustid/chromaprint/src/3a2cd633e0f220295239f0056a34c8ce0d51ce9a/examples/fpcalc.c?at=master modifié le 22/07/14 : suppression des appels de fonction dépréciée
+      //https://bitbucket.org/acoustid/chromaprint/src/3a2cd633e0f220295239f0056a34c8ce0d51ce9a/examples/fpcalc.c?at=master 
+      //modifié le 22/07/14 : suppression des appels de fonction dépréciée
       int ChromaprintFingerprinter::decode_audio_file(ChromaprintContext *chromaprint_ctx, const char *file_name, int max_length, int *duration, String^% error)
 	   {
 		   int ok = 0, remaining, length, consumed, codec_ctx_opened = 0, got_frame, stream_index;
@@ -108,10 +109,10 @@ namespace Luminescence
 		   }
 
 		   if (stream->duration != AV_NOPTS_VALUE) {
-			   *duration = stream->time_base.num * stream->duration / stream->time_base.den;
+			   *duration = (int)(stream->time_base.num * stream->duration / stream->time_base.den);
 		   }
 		   else if (format_ctx->duration != AV_NOPTS_VALUE) {
-			   *duration = format_ctx->duration / AV_TIME_BASE;
+			   *duration = (int)(format_ctx->duration / AV_TIME_BASE);
 		   }
 		   else {
 			   //fprintf(stderr, "ERROR: couldn't detect the audio duration\n");
